server/requirements.txt
face-recognition==1.2.3
opencv-python==3.4.4.19
Pillow==5.3.0
pymongo==3.7.2
tornado==5.1.1
Werkzeug==0.14.1
WTForms==2.2.1
